% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeDates.R
\name{changeDates}
\alias{changeDates}
\title{changeDates}
\usage{
changeDates(idx)
}
\arguments{
\item{date}{a date character vector containing "m" or "q".}
}
\description{
change statlink dates to R Date
}
\details{
change statlink dates to R Date.
}
\examples{
changeDates(c("2012m1", "2012m2"))
changeDates(c("2012q1", "2012q2", "2012q3", "2012q4"))
}
\author{
Bo Werth
}

